+++ app/components/games/PenaltyScoreboard 2.tsx
@@ -1,3 +1,5 @@
+export {}
+
 'use client'
 
 import React, { useEffect, useRef, useState } from 'react'